A complete rewrite of cocoshop, built with the MVC pattern
I decided that cocoshop was a little messy and getting the current
version to work with multiple projects would be a challenge, so I
thought the best way to handle that would be to rewrite it. It still is
a WIP and not ready for use yet, but I'm very pleased with it so far.
In past versions, I had been making my own sidebar for editing the
nodes. I found a new BSD licensed project called TLAnimatingOutlineView,
which allows a dynamic and easy way to make views that are collapsible
and that resize correctly. I've made several changes to it to allow for
larger disclosure views and to make it compile with the LLVM compiler
and to silence some warnings.  Additionally, I'm using PSMTabBarControl
for, well, a tab bar. This is working fine, but it's getting a little
outdated and doesn't look like a modern Cocoa app. I love Chromium's
tabs, and I was planning on porting it to Cocoa, but I found that
someone has already done that ( https://github.com/rsms/chromium-tabs ).
I plan on migrating the current tab bar to the Chromium tab bars.  The
reason that I'm using tabs rather than windows is because I'm not sure
how to have multiple instances of cocos2d at once. It might be possible
if I were to create separate processes for cocoshop and the cocos2d
renderer, but that is a little complicated and for now I'd like to just
keep it how it is. Maybe in the future we could have multiple windows,
which would then allow for side-by-side editing.  Another change I've
made is for the CSNodes. I've taken the route of SpaceManager where I
#define all the methods, variables, etc. and then have a protocol for
the CSNode that all of the custom objects conform to. You can look at
CSSprite to see an example of how CSNodes can easily be implemented.
There are still a few bugs, and it doesn't have an export/import feature
yet, so I plan on working on that.