Added Chrome style tabs
Thanks to rsms's chromium-tabs, I was able to implement the tabs used in
Chromium/Google Chrome for cocoshop. The implementation isn't as
straightforward as they were for PSMTabBarControl, which I'd been using
before, but it was still possible. You can look in the header files of
the Chromium Tab subclasses to see an explanation on how it works.